requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-reversion==3.0.3
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
jieba==0.39
Markdown==3.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.24.3
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
